requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
Pillow>=10.0.0
icalendar>=6.0.0
